Componentes extraídos de CreateEventPage a carpeta components
